Include dependencies of input file.
This uses the dts-bundle-generator package to generate a single d.ts file containing the input file and any dependencies, which we then pass to ts-morph.

Although ts-morph does support finding the dependencies of an input source file, it still includes all declarations within the files, rather than just those that are referenced. By using dts-bundle-generator this is handled for us.

We also update our tests to be more comprehensive and cover this.